requirements.txt
h5py==2.8.0
scipy==1.1.0
requests==2.20.1
torchvision==0.2.1
Cython==0.29.12
numpy==1.13.3
torch==0.4.1.post2
Pillow==7.2.0